chicky_data_pipeline/requirements.txt
wordfreq==3.1.1
nltk==3.9.1
asyncpg==0.29.0
python-dotenv==1.0.1
anthropic==0.34.2
openai==1.51.0
pandas==2.2.2
tqdm==4.66.5
aiohttp==3.10.5
